+++ libqpdf/QPDFFormFieldObjectHelper.cc
@@ -362,12 +362,10 @@ QPDFFormFieldObjectHelper::setV(QPDFObjectHandle value, bool need_appearances)
         setFieldAttribute("/V", value);
     }
     if (need_appearances) {
-        QPDF* qpdf = this->oh.getOwningQPDF(
-            false,
+        QPDF& qpdf = this->oh.getQPDF(
             "QPDFFormFieldObjectHelper::setV called with need_appearances = "
             "true on an object that is not associated with an owning QPDF");
-
-        QPDFAcroFormDocumentHelper(*qpdf).setNeedAppearances(true);
+        QPDFAcroFormDocumentHelper(qpdf).setNeedAppearances(true);
     }
 }
 
@@ -881,7 +879,7 @@ QPDFFormFieldObjectHelper::generateTextAppearance(
         if (found_font_in_dr && resources.isDictionary()) {
             QTC::TC("qpdf", "QPDFFormFieldObjectHelper get font from /DR");
             if (resources.isIndirect()) {
-                resources = resources.getOwningQPDF(false)->makeIndirectObject(
+                resources = resources.getQPDF().makeIndirectObject(
                     resources.shallowCopy());
                 AS.getDict().replaceKey("/Resources", resources);
             }